refactor: subclass explorer form mda widget
